#43 - 리팩토링 : request, response dto 위치 변경
